Microservice_Notifications/requirements.txt
Django==5.2.0
djangorestframework==3.15.0
django-cors-headers==4.0.0
python-dotenv==1.0.0
twilio==8.2.0
mysqlclient==2.2.0
# Optional / useful packages
gunicorn==21.2.0  # production WSGI server (useful when deploying with Docker)